check file existence before removing it
This is an attempt to resolve an issue with warnings being thrown by
hhvm.

+++ src/Stash/Utilities.php
@@ -149,7 +149,9 @@ class Utilities
                     $dirFiles = scandir($file->getPathname());
                     if ($dirFiles && count($dirFiles) == 2) {
                         $filename = rtrim($filename, '/.');
-                        rmdir($filename);
+                        if (file_exists($filename)) {
+                            rmdir($filename);
+                        }
                     }
                     unset($dirFiles);
                     continue;
@@ -157,7 +159,9 @@ class Utilities
 
                 if (file_exists($filename)) {
                     if ($file->isDir()) {
-                        rmdir($filename);
+                        if (file_exists($filename)) {
+                            @rmdir($filename);
+                        }
                     } else {
                         unlink($filename);
                     }
